Extract node controller to own package
