+++ pkg/virt-controller/watch/BUILD.bazel
@@ -6,7 +6,6 @@ go_library(
         "application.go",
         "migration.go",
         "migrationpolicy.go",
-        "node.go",
         "replicaset.go",
         "util.go",
         "vm.go",
@@ -38,7 +37,6 @@ go_library(
         "//pkg/storage/types:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/cluster:go_default_library",
-        "//pkg/util/lookup:go_default_library",
         "//pkg/util/migrations:go_default_library",
         "//pkg/util/pdbs:go_default_library",
         "//pkg/util/ratelimiter:go_default_library",
@@ -52,6 +50,7 @@ go_library(
         "//pkg/virt-controller/watch/common:go_default_library",
         "//pkg/virt-controller/watch/drain/disruptionbudget:go_default_library",
         "//pkg/virt-controller/watch/drain/evacuation:go_default_library",
+        "//pkg/virt-controller/watch/node:go_default_library",
         "//pkg/virt-controller/watch/pool:go_default_library",
         "//pkg/virt-controller/watch/topology:go_default_library",
         "//pkg/virt-controller/watch/util:go_default_library",
@@ -80,7 +79,6 @@ go_library(
         "//vendor/k8s.io/apimachinery/pkg/api/equality:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/fields:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/labels:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/types:go_default_library",
@@ -107,7 +105,6 @@ go_test(
     srcs = [
         "application_test.go",
         "migration_test.go",
-        "node_test.go",
         "replicaset_test.go",
         "vm_test.go",
         "vmi_test.go",
@@ -132,6 +129,7 @@ go_test(
         "//pkg/virt-controller/watch/common:go_default_library",
         "//pkg/virt-controller/watch/drain/disruptionbudget:go_default_library",
         "//pkg/virt-controller/watch/drain/evacuation:go_default_library",
+        "//pkg/virt-controller/watch/node:go_default_library",
         "//pkg/virt-controller/watch/topology:go_default_library",
         "//pkg/virt-controller/watch/util:go_default_library",
         "//staging/src/kubevirt.io/api/clone/v1alpha1:go_default_library",
