+++ pkg/virt-controller/watch/node/node.go
@@ -1,4 +1,4 @@
-package watch
+package node
 
 import (
 	"context"
