+++ pkg/virt-controller/watch/node/node_test.go
@@ -1,4 +1,4 @@
-package watch
+package node
 
 import (
 	"context"
